refactor: 중복되는 Dao 통합
