+++ src/main/java/service/ChessStatisticService.java
@@ -1,7 +1,7 @@
 package service;
 
-import dao.StatusRecordWithRoomNameDao;
-import dto.StatusRecordWithRoomNameDto;
+import dao.StatusRecordDao;
+import dto.StatusRecordDto;
 
 import java.sql.SQLException;
 import java.util.List;
@@ -10,21 +10,21 @@ public class ChessStatisticService {
 	private static final ChessStatisticService STATISTIC_SERVICE;
 
 	static {
-		STATISTIC_SERVICE = new ChessStatisticService(StatusRecordWithRoomNameDao.getInstance());
+		STATISTIC_SERVICE = new ChessStatisticService(StatusRecordDao.getInstance());
 	}
 
-	private final StatusRecordWithRoomNameDao statusRecordWithRoomNameDao;
+	private final StatusRecordDao statusRecordDao;
 
-	public ChessStatisticService(final StatusRecordWithRoomNameDao statusRecordWithRoomNameDao) {
-		this.statusRecordWithRoomNameDao = statusRecordWithRoomNameDao;
+	public ChessStatisticService(final StatusRecordDao statusRecordDao) {
+		this.statusRecordDao = statusRecordDao;
 	}
 
 	public static ChessStatisticService getInstance() {
 		return STATISTIC_SERVICE;
 	}
 
-	public List<StatusRecordWithRoomNameDto> loadStatusRecordsWithRoomName() throws SQLException {
-		return statusRecordWithRoomNameDao.findStatusRecordWithRoomName();
+	public List<StatusRecordDto> loadStatusRecordsWithRoomName() throws SQLException {
+		return statusRecordDao.findStatusRecords();
 	}
 
 }
